Chapter 1 Files. App named, Chapter 1 assumes the reader has chosen to run the diagram command discussed in this chapter for a project that already exists, as in the reader had already worked with or is currently working with Django and wishes to follow along with the instructions in the book for a project that is already running.
The folder "diagrams\chapter_1" is where you can find the diagram that was auto generated. The models Team, and Award were created as examples, the remainer User and other tables all come standard with a basic installation of Django by running the startproject and startapp commands. The user will get a full break down on how to do that in Chapter 2, which if they follow exactly will produce the exact same file structure as is the intent, with each chapter separated by chapter_# apps. This is why the default db.sqlite3 database file exists. Theoretically a reader should be able to just run this as is granted they followed everything in Chapter 1 to set up Pyhon, Django and other tools needed just to run the project on their local machine.